move SamplePedestalEvents to dedicated package
