+++ src/main/java/fact/pedestalSuperposition/SamplePedestalEvent.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package fact.utils;
+package fact.pedestalSuperposition;
 
 import fact.Utils;
 import fact.container.PreviousEventInfoContainer;
