+++ src/main/java/fact/pedestalSuperposition/package-info.java
@@ -1,0 +1,9 @@
+package fact.pedestalSuperpostion;
+
+/**
+ * @author michael bulinski, michael.bulinski@tu-dortmund.de; Jens Buß, jens.buss@tu-dortmund.de
+ * This package contains a processors to sample and superimpose pedestal events with events from
+ * the input stream e.g. shower simulations without noise simulation.
+ *
+ * The processor <li>{@link fact.pedestalSuperpostion.SamplePedestalEvent}
+ **/
